// src/types/routes.ts
export enum MenuPanels {
  Menu = "default",
  Notifications = "notifications",
  Settings = "settings",
  Achievements = "achievements",
  Bank = "bank",
  Job = "job",
  Earnings = "earnings",
  Business = "business",
  Mining = "mining",
  Investments = "investments",
  Entertainment = "entertainment",
  Quests = "quests",
  Present = "present",
  Bonus = "bonus",
  Community = "community",
  Help = "help",
  PremiumStore = "premium_store",
  Auction = "auction",
  Market = "market",
  Shop = "shop",
}